refactor(mespapiers): Extract widgets proptypes to mutualize
