+++ packages/cozy-mespapiers-lib/src/components/ModelSteps/widgets/proptypes.js
@@ -1,0 +1,21 @@
+import PropTypes from 'prop-types'
+
+const attrsProptypes = PropTypes.shape({
+  name: PropTypes.string,
+  inputLabel: PropTypes.string,
+  type: PropTypes.string,
+  required: PropTypes.bool,
+  minLength: PropTypes.number,
+  maxLength: PropTypes.number,
+  mask: PropTypes.string,
+  maskPlaceholder: PropTypes.string
+})
+
+export const defaultProptypes = {
+  attrs: attrsProptypes.isRequired,
+  defaultValue: PropTypes.string,
+  setValue: PropTypes.func.isRequired,
+  setValidInput: PropTypes.func.isRequired,
+  setIsFocus: PropTypes.func.isRequired,
+  idx: PropTypes.number
+}
